+++ Runtime/DiscreteDistributions/BinomialDistribution/Generators/BinomialGeneratorSimple.cs
@@ -14,7 +14,8 @@ namespace Zor.RandomGenerators.DiscreteDistributions
 	public sealed class BinomialGeneratorSimple : IBinomialGenerator
 	{
 #pragma warning disable CS0649
-		[SerializeField, Range(0f, 0.999999881f)] private float m_Probability = BinomialDistribution.DefaultProbability;
+		[SerializeField, Range(0f, NumberConstants.SubOne)]
+		private float m_Probability = BinomialDistribution.DefaultProbability;
 		[SerializeField] private byte m_UpperBound = BinomialDistribution.DefaultUpperBound;
 #pragma warning restore CS0649