+++ Runtime/DiscreteDistributions/NegativeBinomialDistribution/GeneratorProviders/NegativeBinomialGeneratorDependentProvider.cs
@@ -24,7 +24,7 @@ namespace Zor.RandomGenerators.DiscreteDistributions
 		[SerializeField, Tooltip("Random generator that returns an independent and identically distributed random value in range [0, 1].")]
 		private ContinuousGeneratorProviderReference m_DependedGeneratorProvider;
 		[SerializeField] private int m_StartPoint = NegativeBinomialDistribution.DefaultStartPoint;
-		[SerializeField, Range(1.19E-07f, 1f)]
+		[SerializeField, Range(NumberConstants.NormalEpsilon, 1f)]
 		private float m_Probability = NegativeBinomialDistribution.DefaultProbability;
 		[SerializeField, SimpleRangeInt(1, 255)]
 		private byte m_Successes = NegativeBinomialDistribution.DefaultSuccesses;