+++ Runtime/DiscreteDistributions/NegativeBinomialDistribution/Generators/NegativeBinomialGeneratorSimple.cs
@@ -15,7 +15,7 @@ namespace Zor.RandomGenerators.DiscreteDistributions
 	public sealed class NegativeBinomialGeneratorSimple : INegativeBinomialGenerator
 	{
 #pragma warning disable CS0649
-		[SerializeField, Range(1.19E-07f, 1f)]
+		[SerializeField, Range(NumberConstants.NormalEpsilon, 1f)]
 		private float m_Probability = NegativeBinomialDistribution.DefaultProbability;
 		[SerializeField, SimpleRangeInt(1, 255)]
 		private byte m_Successes = NegativeBinomialDistribution.DefaultSuccesses;